+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
  - Relax some dependency minimum versions
  - Add `size_hint()` to some array subset iterators
+ - **Breaking**: Fix `LinearisedIndicesIterator` to use array shape instead of subset shape
+
 
 ## [0.3.0] - 2023-09-27